tests ~ only test examples prior to version changes and during prepublication testing
## [why]

Project examples use the distribution code which needs to be up-to-date for correct
results. This is guaranteed, by scripted lifecycle events, during version updates and
prepublication. Additionally, code coverage is confused and incorrectly perceives the code
base as split between 'build' and 'dist' when examples, which load the 'dist' modules, are
measured.