fix ~ remove `/// <reference path...` references to deno types
## [why]

Deno barfs to a halt with type conflict errors when it's not exactly in sync with the
otherwise referenced types, which is not controllable by the repository. So, remove any
references to Deno types from the TS files. Intellisense is broken by this but the code
will now run.

Both the built-in JS/TS plugin (vscode.typescript-language-features) and the
Deno-specific plugin (denoland.vscode-deno) are currently unable to understand and work
within a mixed repository.

Await better tooling.