docs ~ (README) stabilize formatting against changes by deno and/or dprint formatters
# [why]

This construction of comments will maintain a stable, Prettier-type formatting even if a
reformat of the file is attempted with `deno fmt` or `dprint fmt`.

Notably, comments prior to `<!DOCTYPE...>` are legal in the standard, but may throw the
MS IE browser into "quirks mode". However, when using modern (2020+) markdown renderers,
this *should not* cause any changes to final rendered output.

ref: [Can comments appear before the DOCTYPE declaration?](https://stackoverflow.com/questions/941100/can-comments-appear-before-the-doctype-declaration)